Начал поправлять ошибки: вынес повыше роутер и передал ему в аргументы нормальную компоненту, подправил ts типы, начал использовать контекст, пофиксил ошибку рендера при нулевом количестве комментов, вынес все интерфейсы в отдельный файл чтобы разбить цикличный импорт, интерполировал текст в html каркас для комментариев, переписал обновление новостей так что теперь отрисовка идет после получения инфы о всех ста новостях, удалил лишние папки и лишнюю вложенность элементов
