+++ src/App.tsx
@@ -1,12 +1,12 @@
-import "./App.css";
-import HackerNews from "./HackerNews/HackerNews";
-import React from "react";
+import './App.css';
+import HackerNews from './HackerNews/HackerNews';
+import React from 'react';
 
 const App: () => JSX.Element = () => {
   return (
-    <div className="App">
+    <>
       <HackerNews />
-    </div>
+    </>
   );
 };
